refactor: consolidate errorMessage / truncate into @mulmoclaude/core (#2217)
`errorMessage` existed 14 times across 4 behaviours; six of those copies sat
inside packages/core alone, where nothing required them — intra-package imports
are unrestricted. The behaviours disagreed on gRPC-shaped errors: `{ code,
details }` surfaced as "quota exceeded" through the host copy and as
"[object Object]" through the core ones. `truncate`'s core copy had dropped the
`ellipsis.length >= max` guard, so it returned 3 chars for
`truncate("hello", 2, "...")` while its own docblock promised output never
exceeds `max`.

None of these were firing in production — every caller catches Error instances,
where all variants agree, and the sole core `truncate` caller uses the 1-char
default ellipsis. This removes the traps, not a live bug.

Core cannot import server/, so the fix is the "pull it OUT into core" pattern:
canonical lives in a new browser-safe `@mulmoclaude/core/utils`, and the host's
three helper modules re-export it. Re-export rather than repointing 92 files /
342 call sites — the goal is one implementation, and the reason is written into
each file.

Google's `"unknown error"` default is dropped rather than threaded through as a
fallback: measuring showed the canonical returns `fallback` before `String(err)`,
so a thrown string came back as "unknown error" instead of its own content. It
also recreated a second behaviour, which is what this change removes.

Version bump is load-bearing: published core@0.26.0 has no `./utils` subpath, so
the npm-installed launcher would fail to resolve the host's import. core →
0.27.0, launcher range and the two plugin peer ranges follow. **core@0.27.0 must
be published before the launcher is published next.**

docs/shared-utils.md reviewed end to end: six rotted entries corrected (three
pointed at deleted files), and a "Known duplicates" table added — the catalog's
failure mode was naming one member of a family and hiding the rest, which is how
this grew to 14 in the first place.
