+++ packages/core/src/utils/index.ts
@@ -1,0 +1,4 @@
+// Entry point for `@mulmoclaude/core/utils`. Follows the one-index-per-subpath
+// convention the other core entries use (see vite.config.ts's entry map).
+export { errorMessage, toError } from "./errors.js";
+export { truncate } from "./text.js";
