+++ packages/mulmoclaude/package.json
@@ -37,7 +37,7 @@
     "@mulmoclaude/accounting-plugin": "^0.3.2",
     "@mulmoclaude/chart-plugin": "^0.1.3",
     "@mulmoclaude/collection-plugin": "^0.13.2",
-    "@mulmoclaude/core": "^0.26.0",
+    "@mulmoclaude/core": "^0.27.0",
     "@mulmoclaude/form-plugin": "^0.1.4",
     "@mulmoclaude/google-plugin": "^0.3.1",
     "@mulmoclaude/html-plugin": "^0.2.5",
