+++ packages/plugins/google-plugin/package.json
@@ -21,7 +21,7 @@
     "test": "tsx --test test/test_*.ts"
   },
   "dependencies": {
-    "@mulmoclaude/core": "^0.26.0"
+    "@mulmoclaude/core": "^0.27.0"
   },
   "peerDependencies": {
     "gui-chat-protocol": "^0.4.0",
